# web/dashboard/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing LoRaWAN rain barrel devices and sensor data",
        "title": "Regenfass API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:64000",
    "basePath": "/",
    "paths": {
        "/device": {
            "post": {
                "description": "Register a device using it's device EUI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Register a device",
                "parameters": [
                    {
                        "description": "Device to register",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterDevicePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.RegisterDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Device EUI is already registered",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}": {
            "get": {
                "description": "Retrieve device details using either read-write or read-only token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Get device information by token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device token (read-write or read-only)",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeviceInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update mutable device fields (currently only the custom name). Requires the read-write token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Update device fields",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device read-write token",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateDevicePayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Device token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}/channels/{channelID}": {
            "put": {
                "description": "Set the name and declared measurement type of one of a device's channels. The channel does not have to have reported anything yet — describing it up front is how a slot is prepared for a sensor. Requires the read-write token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Describe a device channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device read-write token",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Channel id (0–15)",
                        "name": "channelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "How to describe the channel",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertDeviceChannelPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token, channel id or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Device token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}/channels/{channelID}/hidden": {
            "put": {
                "description": "Takes a channel off the dashboard, or puts it back. Nothing is deleted: the channel's measurements stay and new ones keep arriving, they are simply not returned by the measurement endpoints while it is hidden. A hidden channel still appears in a device's `channels` so it can be restored. Requires the read-write token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Hide or restore a device channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device read-write token",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Channel id (0–15)",
                        "name": "channelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Whether to hide the channel",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SetDeviceChannelHiddenPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token, channel id or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Device token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}/measurements": {
            "get": {
                "description": "Retrieve a device's measurement history between `start` and `end`, downsampled to a resolution that scales with the span. The endpoint aims for roughly 2000 data points per channel: wider ranges are bucketed more coarsely (never coarser than three points per day) and narrower ranges more finely (never finer than one point per minute). Each bucket is represented by its newest reading. Optionally restrict to a single channel. Authenticated by either the read-write or read-only device token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Get downsampled device measurements over a time range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device token (read-write or read-only)",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start of the range, inclusive (RFC3339)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of the range, inclusive (RFC3339)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Only return measurements for this channel id",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.RangedMeasurementsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token or query parameter",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/group": {
            "post": {
                "description": "Create a new device group for organizing rain barrel devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create a new group",
                "parameters": [
                    {
                        "description": "Group name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateGroupPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.CreateGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/group/{groupToken}": {
            "get": {
                "description": "Retrieve group details and associated devices using either read-write or read-only token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get group information by token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group token (read-write or read-only)",
                        "name": "groupToken",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GroupInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid group token",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/group/{groupToken}/devices": {
            "post": {
                "description": "Add an existing device to a group using the group's read-write token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Add device to group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group read-write token",
                        "name": "groupToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device token to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddDeviceToGroupPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Group token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/ingest": {
            "post": {
                "description": "Process LoRaWAN uplink messages and store device measurements",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Ingest LoRaWAN uplink data",
                "parameters": [
                    {
                        "description": "LoRaWAN uplink message",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IngestRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/measurements/latest": {
            "post": {
                "description": "Returns one entry per device referenced (directly via `devices` or transitively via `groups`), including its name, location, and the most recent measurement per channel. Devices with no measurements yet still appear with an empty `measurements` list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements"
                ],
                "summary": "Get latest measurements grouped per device",
                "parameters": [
                    {
                        "description": "Tokens selecting devices and/or groups",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LatestMeasurementsPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.LatestMeasurementsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/overview": {
            "post": {
                "description": "Returns the dashboard view for a set of subscription tokens: every referenced group with its member devices nested inside, plus every directly-referenced device in `devices`. A device that is both a group member and subscribed directly appears in both places. Each device carries its latest reading per channel, exactly as in `/measurements/latest`. Unlike that endpoint, group membership is preserved so the UI can render devices under their groups.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements"
                ],
                "summary": "Get subscriptions organized by group",
                "parameters": [
                    {
                        "description": "Tokens selecting groups and/or devices",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LatestMeasurementsPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OverviewResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.CreateGroupPayload": {
            "description": "Request body to create a new device group",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "My Rain Barrel Group"
                }
            }
        },
        "api.CreateGroupResponse": {
            "description": "Response with tokens",
            "type": "object",
            "properties": {
                "read_only_token": {
                    "type": "string",
                    "example": "ro_group_token_123"
                },
                "read_write_token": {
                    "type": "string",
                    "example": "rw_group_token_123"
                }
            }
        },
...
        "api.DeviceInfoResponse": {
            "description": "Identity of a device. When authenticated with the read-write token, both tokens are returned; otherwise only the read-only token is returned. `name` is the user-set name, or a stable auto-generated nickname if none has been set.",
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_readonly": {
                    "type": "boolean",
                    "example": false
                },
                "latitude": {
                    "type": "number",
                    "example": 51.3397
                },
                "longitude": {
                    "type": "number",
                    "example": 12.3731
                },
                "name": {
                    "type": "string",
                    "example": "Happy Barrel"
                },
                "read_only_token": {
                    "type": "string",
                    "example": "ro_token_123"
                },
                "read_write_token": {
                    "type": "string",
                    "example": "rw_token_123"
                }
            }
        },
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "api.RangedMeasurementsResponse": {
            "description": "Downsampled measurement series for a device over a time range. The resolution adapts to the span: the wider the range, the coarser the buckets (never finer than one point per minute, never coarser than three points per day). `bucket_seconds` reports the bucket width that was actually used.",
            "type": "object",
            "properties": {
                "bucket_seconds": {
                    "type": "number",
                    "example": 302.4
                },
                "channel_id": {
                    "type": "integer",
                    "example": 1
                },
                "end": {
                    "type": "string",
                    "example": "2024-01-08T00:00:00Z"
                },
                "measurements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.DeviceMeasurement"
                    }
                },
                "start": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "api.RegisterDevicePayload": {
            "type": "object",
            "properties": {
                "device_eui": {
                    "type": "string",
                    "example": "AABBCCDDEEFF0011"
                }
            }
        },
        "api.RegisterDeviceResponse": {
            "type": "object",
            "properties": {
                "read_only_token": {
                    "type": "string",
                    "example": "ro_token_123"
                },
                "read_write_token": {
                    "type": "string",
                    "example": "rw_token_123"
                }
            }
        },
        "api.SetDeviceChannelHiddenPayload": {
            "description": "Whether the channel should be shown on the dashboard.",
            "type": "object",
            "properties": {
                "hidden": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "api.UpdateDevicePayload": {
            "description": "Fields that can be updated on a device (only present fields are applied)",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Rain barrel by the shed"
                }
            }
        },
        "api.UpsertDeviceChannelPayload": {
            "description": "How a device's channel should be described. `measurement_type` is the type the user picked in the dashboard. Send either field as null (or the name as empty) to clear it, which returns the channel to being undescribed.",
            "type": "object",
            "properties": {
                "measurement_type": {
                    "type": "integer",
                    "example": 4
                },
                "name": {
                    "type": "string",
                    "example": "Water Level"
                }
            }
        }
    }
}